requirements.txt
astroid
isort
lazy-object-proxy
mccabe
pygame
pylint
six
typed-ast
wrapt
tensorflow
numpy
pandas
neat-python
